+++ src/document_manager.cpp
@@ -113,7 +113,7 @@ void document_manager::close_document(int index) {
 	}
 	const document_tab* tab = get_tab(index);
 	if (tab != nullptr && tab->text_ctrl != nullptr) {
-		const long position = tab->text_ctrl->GetInsertionPoint();
+		const int position = tab->text_ctrl->GetInsertionPoint();
 		save_document_position(tab->file_path, position);
 		config.set_document_opened(tab->file_path, false);
 	}
@@ -181,13 +181,13 @@ int document_manager::get_active_tab_index() const {
 	return notebook->GetSelection();
 }
 
-void document_manager::go_to_position(long position) const {
+void document_manager::go_to_position(int position) const {
 	wxTextCtrl* text_ctrl = get_active_text_ctrl();
 	if (text_ctrl == nullptr) {
 		return;
 	}
-	const long max_pos = text_ctrl->GetLastPosition();
-	position = std::clamp(position, 0L, max_pos);
+	const int max_pos = text_ctrl->GetLastPosition();
+	position = std::clamp(position, 0, max_pos);
 	text_ctrl->SetInsertionPoint(position);
 	text_ctrl->ShowPosition(position);
 }
@@ -207,7 +207,7 @@ void document_manager::go_to_previous_section() const {
 	bool wrapping = false;
 	const int current_index = doc->section_index(current_pos);
 	if (current_index != -1) {
-		const size_t current_section_offset = doc->offset_for_section(current_index);
+		const int current_section_offset = doc->offset_for_section(current_index);
 		if (current_pos > current_section_offset) {
 			text_ctrl->SetInsertionPoint(static_cast<long>(current_section_offset));
 			long line{0};
@@ -219,7 +219,7 @@ void document_manager::go_to_previous_section() const {
 	}
 	int search_pos = current_pos;
 	if (current_index != -1) {
-		const size_t current_section_offset = doc->offset_for_section(current_index);
+		const int current_section_offset = doc->offset_for_section(current_index);
 		if (current_pos <= current_section_offset) {
 			// We're at the start of the current section, so search from just before the section marker.
 			search_pos = current_section_offset > 0 ? current_section_offset - 1 : 0;
@@ -238,7 +238,7 @@ void document_manager::go_to_previous_section() const {
 			return;
 		}
 	}
-	const size_t offset = doc->offset_for_section(prev_index);
+	const int offset = doc->offset_for_section(prev_index);
 	text_ctrl->SetInsertionPoint(static_cast<long>(offset));
 	long line{0};
 	text_ctrl->PositionToXY(static_cast<long>(offset), nullptr, &line);
@@ -276,7 +276,7 @@ void document_manager::go_to_next_section() const {
 			return;
 		}
 	}
-	const size_t offset = doc->offset_for_section(next_index);
+	const int offset = doc->offset_for_section(next_index);
 	text_ctrl->SetInsertionPoint(static_cast<long>(offset));
 	long line{0};
 	text_ctrl->PositionToXY(static_cast<long>(offset), nullptr, &line);
@@ -329,7 +329,7 @@ void document_manager::go_to_previous_page() const {
 			return;
 		}
 	}
-	const size_t offset = doc->offset_for_page(prev_index);
+	const int offset = doc->offset_for_page(prev_index);
 	text_ctrl->SetInsertionPoint(static_cast<long>(offset));
 	long line{0};
 	text_ctrl->PositionToXY(static_cast<long>(offset), nullptr, &line);
@@ -367,7 +367,7 @@ void document_manager::go_to_next_page() const {
 			return;
 		}
 	}
-	const size_t offset = doc->offset_for_page(next_index);
+	const int offset = doc->offset_for_page(next_index);
 	text_ctrl->SetInsertionPoint(static_cast<long>(offset));
 	long line{0};
 	text_ctrl->PositionToXY(static_cast<long>(offset), nullptr, &line);
@@ -558,7 +558,7 @@ void document_manager::activate_current_link() const {
 	if (doc == nullptr || text_ctrl == nullptr) {
 		return;
 	}
-	const size_t current_pos = text_ctrl->GetInsertionPoint();
+	const int current_pos = text_ctrl->GetInsertionPoint();
 	const int link_index = doc->buffer.current_marker_index(current_pos, marker_type::link);
 	if (link_index == -1) {
 		return;
@@ -631,12 +631,12 @@ void document_manager::toggle_bookmark() {
 	long selection_start = 0;
 	long selection_end = 0;
 	text_ctrl->GetSelection(&selection_start, &selection_end);
-	long bookmark_start, bookmark_end;
+	int bookmark_start, bookmark_end;
 	if (selection_end > selection_start) {
-		bookmark_start = selection_start;
-		bookmark_end = selection_end;
+		bookmark_start = static_cast<int>(selection_start);
+		bookmark_end = static_cast<int>(selection_end);
 	} else {
-		const long current_pos = text_ctrl->GetInsertionPoint();
+		const int current_pos = text_ctrl->GetInsertionPoint();
 		bookmark_start = current_pos;
 		bookmark_end = current_pos;
 	}
@@ -663,12 +663,12 @@ void document_manager::add_bookmark_with_note() {
 	long selection_start{0};
 	long selection_end{0};
 	text_ctrl->GetSelection(&selection_start, &selection_end);
-	long bookmark_start, bookmark_end;
+	int bookmark_start, bookmark_end;
 	if (selection_end > selection_start) {
-		bookmark_start = selection_start;
-		bookmark_end = selection_end;
+		bookmark_start = static_cast<int>(selection_start);
+		bookmark_end = static_cast<int>(selection_end);
 	} else {
-		const long current_pos = text_ctrl->GetInsertionPoint();
+		const int current_pos = text_ctrl->GetInsertionPoint();
 		bookmark_start = current_pos;
 		bookmark_end = current_pos;
 	}
@@ -710,13 +710,13 @@ void document_manager::show_bookmark_dialog(wxWindow* parent) {
 		speak(_("No bookmarks"));
 		return;
 	}
-	const long current_pos = text_ctrl->GetInsertionPoint();
+	const int current_pos = text_ctrl->GetInsertionPoint();
 	bookmark_dialog dialog(parent, bookmarks, text_ctrl, config, tab->file_path, current_pos);
 	const int result = dialog.ShowModal();
 	if (result != wxID_OK) {
 		return;
 	}
-	const long pos = dialog.get_selected_position();
+	const int pos = dialog.get_selected_position();
 	if (pos < 0) {
 		return;
 	}
@@ -762,7 +762,7 @@ void document_manager::show_table_of_contents(wxWindow* parent) const {
 		speak(_("Table of contents is empty."));
 		return;
 	}
-	const size_t current_pos = text_ctrl->GetInsertionPoint();
+	const int current_pos = text_ctrl->GetInsertionPoint();
 	const int closest_toc_offset = doc->find_closest_toc_offset(current_pos);
 	toc_dialog dlg(parent, doc, closest_toc_offset);
 	if (dlg.ShowModal() != wxID_OK) {
@@ -784,12 +784,12 @@ void document_manager::show_document_info(wxWindow* parent) const {
 	dlg.ShowModal();
 }
 
-void document_manager::save_document_position(const wxString& path, long position) const {
+void document_manager::save_document_position(const wxString& path, int position) const {
 	config.set_document_position(path, position);
 	config.flush();
 }
 
-long document_manager::load_document_position(const wxString& path) const {
+int document_manager::load_document_position(const wxString& path) const {
 	return config.get_document_position(path);
 }
 
@@ -798,7 +798,7 @@ void document_manager::save_current_tab_position() const {
 	if (tab == nullptr || tab->text_ctrl == nullptr) {
 		return;
 	}
-	const long position = tab->text_ctrl->GetInsertionPoint();
+	const int position = tab->text_ctrl->GetInsertionPoint();
 	save_document_position(tab->file_path, position);
 }
 
@@ -806,7 +806,7 @@ void document_manager::save_all_tab_positions() const {
 	for (int i = 0; i < get_tab_count(); ++i) {
 		const document_tab* tab = get_tab(i);
 		if (tab != nullptr && tab->text_ctrl != nullptr) {
-			const long position = tab->text_ctrl->GetInsertionPoint();
+			const int position = tab->text_ctrl->GetInsertionPoint();
 			save_document_position(tab->file_path, position);
 		}
 	}
@@ -820,8 +820,8 @@ wxString document_manager::get_status_text() const {
 	if (text_ctrl == nullptr) {
 		return _("Ready");
 	}
-	const long current_pos = text_ctrl->GetInsertionPoint();
-	const long total_chars = text_ctrl->GetLastPosition();
+	const int current_pos = text_ctrl->GetInsertionPoint();
+	const int total_chars = text_ctrl->GetLastPosition();
 	const int percentage = total_chars > 0 ? (current_pos * 100) / total_chars : 0;
 	long line{0};
 	text_ctrl->PositionToXY(current_pos, nullptr, &line);
@@ -838,7 +838,7 @@ wxString document_manager::get_window_title(const wxString& app_name) const {
 	return doc != nullptr ? app_name + " - " + doc->title : app_name;
 }
 
-long document_manager::find_text(const wxString& query, long start_pos, find_options options) const {
+int document_manager::find_text(const wxString& query, int start_pos, find_options options) const {
 	const wxTextCtrl* text_ctrl = get_active_text_ctrl();
 	if (text_ctrl == nullptr) {
 		return wxNOT_FOUND;
@@ -852,7 +852,7 @@ void document_manager::apply_word_wrap(bool word_wrap) {
 		document_tab* tab = get_tab(i);
 		if (tab != nullptr && tab->text_ctrl != nullptr && tab->panel != nullptr) {
 			wxTextCtrl* old_ctrl = tab->text_ctrl;
-			const long current_pos = old_ctrl->GetInsertionPoint();
+			const int current_pos = old_ctrl->GetInsertionPoint();
 			const wxString content = old_ctrl->GetValue();
 			wxSizer* sizer = tab->panel->GetSizer();
 			sizer->Detach(old_ctrl);
@@ -910,9 +910,9 @@ void document_manager::restore_document_position(document_tab* tab) const {
 	if (tab == nullptr || tab->text_ctrl == nullptr) {
 		return;
 	}
-	const long saved_position = load_document_position(tab->file_path);
+	const int saved_position = load_document_position(tab->file_path);
 	if (saved_position > 0) {
-		const long max_position = tab->text_ctrl->GetLastPosition();
+		const int max_position = tab->text_ctrl->GetLastPosition();
 		if (saved_position <= max_position) {
 			tab->text_ctrl->SetInsertionPoint(saved_position);
 			tab->text_ctrl->ShowPosition(saved_position);
@@ -963,7 +963,7 @@ void document_manager::navigate_to_heading(bool next, int specific_level) const 
 			return;
 		}
 	}
-	const size_t offset = doc->offset_for_heading(target_index);
+	const int offset = doc->offset_for_heading(target_index);
 	text_ctrl->SetInsertionPoint(static_cast<long>(offset));
 	const marker* heading_marker = doc->get_heading_marker(target_index);
 	if (heading_marker != nullptr) {
